Fix #192. (#200)
* use the worker id to increment the ports

* ensure session scoped fixtures are executed once

* update e2e test readme

* Add stub file. Remove ULimit hardcoding.

Also minor style edits.

* remove hard coded port values

* update sleep duration to avoid race conditions

* Add stub file, minor style edits

* give the Meros deamon time to start up with a timeout

* Cleanup.

Said cleanup fixes the ULimit test which wasn't working. My guess is a 
socket was left hanging somewhere which offsetted it.

* Satisfy linter.

+++ e2e/Tests/conftest.py
@@ -1,9 +1,10 @@
 #pylint: disable=invalid-name
 
 from typing import Any
-from time import sleep
+from pathlib import Path
 import shutil
 
+from filelock import FileLock
 from pytest import fixture
 
 from e2e.Meros.Meros import Meros
@@ -12,28 +13,39 @@ from e2e.Meros.RPC import RPC
 #Delete the existing data directory.
 @fixture(scope="session", params=["./data/e2e"])
 def dataDir(
-  request: Any
+  request: Any,
+  tmp_path_factory: Any
 ) -> str:
-  try:
-    shutil.rmtree(request.param)
-  except FileNotFoundError:
-    pass
+  #Get the temp directory shared by all workers.
+  tmpDir: Path = tmp_path_factory.getbasetemp().parent
+  fn: Path = tmpDir / ".clean"
+  with FileLock(str(fn) + ".lock"):
+    if not fn.is_file():
+      try:
+        shutil.rmtree(request.param)
+      except FileNotFoundError:
+        pass
+      fn.write_text("")
   return request.param
 
 @fixture(scope="module", params=[5132])
 def meros(
   #pylint: disable=redefined-outer-name
   dataDir: str,
-  request: Any
+  request: Any,
+  worker_id: str
 ) -> Meros:
+  #If xdist is disabled, the worker_id will return "master" and
+  #"gw1", "gw2", ... otherwise
+  index: int = 0
+  if worker_id.startswith("gw"):
+    index = int(worker_id[2:])
   result: Meros = Meros(
     request.node.module.__name__,
-    request.param,
-    request.param + 1,
+    request.param + (2 * index),
+    request.param + (2 * index) + 1,
     dataDir
   )
-  #Let the instance start up.
-  sleep(5)
   request.addfinalizer(result.quit)
   return result
